reorganize our code base into packages
