+++ ufo/handlers/__init__.py
@@ -1,0 +1,1 @@
+# This makes the tests discoverable in the sub-package.
