+++ ufo/services/key_distributor_test.py
@@ -4,9 +4,9 @@ from mock import patch
 
 import ufo
 from ufo import base_test
-from ufo import key_distributor
-from ufo import models
-from ufo import user_test
+from ufo.database import models
+from ufo.handlers import user_test
+from ufo.services import key_distributor
 
 
 class KeyDistributorTest(base_test.BaseTest):
